Moved user role models to same model file
