--- models/parent.ts
@@ -1,13 +1,0 @@
-import type { province, locale } from "@prisma/client";
-
-export type ParentInput = {
-    id: string;
-    phoneNumber: string;
-    isLowIncome?: boolean;
-    addressLine1: string;
-    addressLine2?: string;
-    postalCode: string;
-    cityName: string;
-    province: province;
-    preferredLanguage: locale;
-};
