--- models/programadmin.ts
@@ -1,3 +1,0 @@
-export type ProgramAdminInput = {
-    id: string;
-};
