--- models/teacher.ts
@@ -1,4 +1,0 @@
-// TODO: not for MVP
-export type TeacherInput = {
-    id: string;
-};
