--- models/volunteer.ts
@@ -1,14 +1,0 @@
-import type { province, locale } from "@prisma/client";
-
-export type VolunteerInput = {
-    id: string;
-    phoneNumber?: string;
-    isValid?: boolean;
-    backgroundFormLink?: string;
-    addressLine1?: string;
-    addressLine2?: string;
-    postalCode?: string;
-    cityName?: string;
-    province?: province;
-    preferredLanguage?: locale;
-};
